Use semver data for equal and compare traits
`Tag` is supposed to use the data in its `semver` attribute for equal
and compare operations since strings like "GE-Proton7-7" and
"6.20-GE-1" can not be easily compared.

However, this has been poorly implemented by providing a method
`cmp_semver` on `Tag` making it possible to not use semver comparison
and instead comparing the `Tag` string values.

To properly fix this the `PartialEq` and `PartialOrd` implementations
for `String` and `str` have been removed.